Demo moved to demo package
